Move load_file_in_db into own python package
